# app/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tap API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/teams": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "List Teams",
        "operationId": "list_teams_api_teams_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TeamResponse"
                  },
                  "type": "array",
                  "title": "Response List Teams Api Teams Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "teams"
        ],
        "summary": "Create Team",
        "operationId": "create_team_api_teams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/teams/{team_id}": {
      "delete": {
        "tags": [
          "teams"
        ],
        "summary": "Delete Team",
        "operationId": "delete_team_api_teams__team_id__delete",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions": {
      "get": {
        "tags": [
          "positions"
        ],
        "summary": "List Positions",
        "operationId": "list_positions_api_positions_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPositions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "positions"
        ],
        "summary": "Create Position",
        "operationId": "create_position_api_positions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions/{position_id}": {
      "get": {
        "tags": [
          "positions"
        ],
        "summary": "Get Position",
        "operationId": "get_position_api_positions__position_id__get",
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "positions"
        ],
        "summary": "Update Position",
        "operationId": "update_position_api_positions__position_id__patch",
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/positions/{position_id}/archive": {
      "post": {
        "tags": [
          "positions"
        ],
        "summary": "Archive Position",
        "operationId": "archive_position_api_positions__position_id__archive_post",
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/candidates/{candidate_id}/positions": {
      "post": {
        "tags": [
          "candidates"
        ],
        "summary": "Add To Position",
        "operationId": "add_to_position_api_candidates__candidate_id__positions_post",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidatePositionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatePositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/positions/{position_id}": {
      "delete": {
        "tags": [
          "candidates"
        ],
        "summary": "Remove From Position",
        "operationId": "remove_from_position_api_candidates__candidate_id__positions__position_id__delete",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          },
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "candidates"
        ],
        "summary": "Update Stage",
        "description": "Update pipeline stage. Return 404 if not found, 422 if invalid transition.",
        "operationId": "update_stage_api_candidates__candidate_id__positions__position_id__patch",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Candidate Id"
            }
          },
          {
            "name": "position_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Position Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatePositionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidatePositionCreate": {
        "properties": {
          "position_id": {
            "type": "integer",
            "title": "Position Id"
          }
        },
        "type": "object",
        "required": [
          "position_id"
        ],
        "title": "CandidatePositionCreate"
      },
      "CandidatePositionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "position_id": {
            "type": "integer",
            "title": "Position Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "candidate_id",
          "position_id",
          "stage",
          "created_at",
          "updated_at"
        ],
        "title": "CandidatePositionResponse"
      },
...
      "CandidateStageItem": {
        "properties": {
          "candidate_id": {
            "type": "integer",
            "title": "Candidate Id"
          },
          "candidate_name": {
            "type": "string",
            "title": "Candidate Name"
          },
          "candidate_email": {
            "type": "string",
            "title": "Candidate Email"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          }
        },
        "type": "object",
        "required": [
          "candidate_id",
          "candidate_name",
          "candidate_email",
          "stage"
        ],
        "title": "CandidateStageItem"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedPositions": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PositionListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "PaginatedPositions"
      },
...
      "PositionCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "requirements": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirements"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "hiring_manager_id": {
            "type": "integer",
            "title": "Hiring Manager Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "team_id",
          "hiring_manager_id"
        ],
        "title": "PositionCreate"
      },
      "PositionDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "requirements": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirements"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "team_name": {
            "type": "string",
            "title": "Team Name"
          },
          "hiring_manager_id": {
            "type": "integer",
            "title": "Hiring Manager Id"
          },
          "hiring_manager_name": {
            "type": "string",
            "title": "Hiring Manager Name"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateStageItem"
            },
            "type": "array",
            "title": "Candidates"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "requirements",
          "status",
          "team_id",
          "team_name",
          "hiring_manager_id",
          "hiring_manager_name",
          "is_archived",
          "candidates",
          "created_at",
          "updated_at"
        ],
        "title": "PositionDetailResponse"
      },
      "PositionListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "team_name": {
            "type": "string",
            "title": "Team Name"
          },
          "hiring_manager_name": {
            "type": "string",
            "title": "Hiring Manager Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "team_name",
          "hiring_manager_name",
          "status",
          "candidate_count"
        ],
        "title": "PositionListItem"
      },
      "PositionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "requirements": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirements"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "hiring_manager_id": {
            "type": "integer",
            "title": "Hiring Manager Id"
          },
          "is_archived": {
            "type": "boolean",
            "title": "Is Archived"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "requirements",
          "status",
          "team_id",
          "hiring_manager_id",
          "is_archived"
        ],
        "title": "PositionResponse"
      },
...
      "PositionUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "requirements": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirements"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "hiring_manager_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hiring Manager Id"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "PositionUpdate"
      },
...
      "StageUpdate": {
        "properties": {
          "stage": {
            "type": "string",
            "title": "Stage"
          }
        },
        "type": "object",
        "required": [
          "stage"
        ],
        "title": "StageUpdate"
      },
...
      "TeamCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "TeamCreate"
      },
      "TeamResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "TeamResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}